Conflicts:
	.gitignore
	SgfTestFiles/19/092.sgf
	SgfTestFiles/9/game-VI-09.sgf
	SgfTestFiles/9/game-VI-12.sgf
	src/orego/mcts/MctsPlayer.java
	src/orego/ui/Orego.java

+++ src/orego/play/Playable.java
@@ -26,7 +26,7 @@ public interface Playable {
 	/**
 	 * @see orego.core.Board#finalScore()
 	 */
-	public int finalScore();
+	public double finalScore();
 
 	/**
 	 * @return the board
